requirements.txt
asgiref==3.8.1
Django==5.0.3
django-environ==0.11.2
djangorestframework==3.15.1
sqlparse==0.4.4
